--- src/view/movies-list-container.js
@@ -1,13 +1,0 @@
-import AbstractView from './abstract.js';
-
-const createMoviesListContainerTemplate = () => (
-  `<div class="films-list__container">
-  </div>
-  `
-);
-
-export default class MoviesListContainerView extends AbstractView {
-  get template() {
-    return createMoviesListContainerTemplate();
-  }
-}
