--- src/view/movies-list.js
@@ -1,14 +1,0 @@
-import AbstractView from './abstract.js';
-
-const createMoviesListTemplate = () => (
-  `<section class="films-list">
-    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-  </section>
-  `
-);
-
-export default class MoviesListView extends AbstractView {
-  get template() {
-    return createMoviesListTemplate();
-  }
-}
